Update wait for write as well
Had to update the wait time for write because I forgot a very important detail:
When you do a write it returns immediately but the I2C bus must still transfer all
of the bytes. It takes about 20ms to send 256 bytes over I2C at 100k baud. Afterward
you must wait 7ms to write the EEPROM page. So a wait of 30ms is best for safety.